final changes by co-author before withdrawal
From Wolfgang:

Alright, let’s fork it.  I’ve attached a final patch for the SRFI 249
document.  It deletes the restarter example (which was oversimplified
and might confuse future readers) as well as a reference to a
“return-denominator-restarter” example which was never written.  For
posterity I’ve also added a link in the Acknowledgments to Taylor
Campbell’s original restarters proposal.